Port TestIcebergPartitionEvolution to JUnit
